backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
shap
python-dotenv
pydantic
reportlab